Guard against potentially unsafe property get
The functions all require the param to be a string, but in real life people will probably pass inn undefined/null in error or edge cases. The lib should return a sane value without causing a runtime error.

This change adds optional chaining to potentially unsafe property gets (.length and .trim) on the elevenDigits value